final update
i had to change the and delete the layout of a some of the buttons to match what i planned. Some challenges i faced, include; changing the colour code to the name of the colour, allowing the photo behind to display in when the cursor hovers over it, making some of the buttons have independent colours.